CLUSTERMON-2205 Refactored cluster-collector operator code
Refactored the code to improve Readability and maintainability aspect of cluster-collector operator code
Created a new controller interface which can be implemented for implementation of collector controllers
