+++ pkg/controller/clustercollector/hostcollector_controller.go
@@ -1,9 +1,11 @@
-package hostcollector
+package clustercollector
 
 import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"strconv"
+	"strings"
 	"time"
 
 	appdynamicsv1alpha1 "github.com/Appdynamics/appdynamics-operator/pkg/apis/appdynamics/v1alpha1"
@@ -27,11 +29,16 @@ import (
 
 const (
 	OLD_SPEC string = "host-collector-spec"
+	HOST_COLLECTOR_NAME string = "k8s-host-collector"
+	CONTAINER_COLLECTOR string = "Container Monitor"
+	TYPE_COLLECTOR string = "Collector"
+	CONTAINER_COLLECTOR_PATH string = "./collectors/containermon-collector-linux-amd64"
+	CONTAINER_CONFIG_NAME string = "container-collector-config"
+	INFRA_AGENT_NAME string = "Infra Structure Agent"
+	INFRA_AGENT_CONFIG_NAME string = "infra-agent-config"
 )
 
 var log = logf.Log.WithName("controller_hostcollector")
-
-const ()
 
 // Add creates a new Hostcollector Controller and adds it to the Manager. The Manager will set fields on the Controller
 // and Start it when the Manager is Started.
@@ -53,7 +60,7 @@ func add(mgr manager.Manager, r reconcile.Reconciler) error {
 	}
 
 	// Watch for changes to primary resource Hostcollector
-	err = c.Watch(&source.Kind{Type: &appdynamicsv1alpha1.Hostcollector{}}, &handler.EnqueueRequestForObject{})
+	err = c.Watch(&source.Kind{Type: &appdynamicsv1alpha1.Clustercollector{}}, &handler.EnqueueRequestForObject{})
 	if err != nil {
 		return err
 	}
@@ -61,7 +68,7 @@ func add(mgr manager.Manager, r reconcile.Reconciler) error {
 	// Watch for changes to secondary resource DaemonSet and requeue the owner Hostcollector
 	err = c.Watch(&source.Kind{Type: &appsv1.DaemonSet{}}, &handler.EnqueueRequestForOwner{
 		IsController: true,
-		OwnerType:    &appdynamicsv1alpha1.Hostcollector{},
+		OwnerType:    &appdynamicsv1alpha1.Clustercollector{},
 	})
 	if err != nil {
 		return err
@@ -84,9 +91,9 @@ func (r *ReconcileHostcollector) Reconcile(request reconcile.Request) (reconcile
 	reqLogger := log.WithValues("Request.Namespace", request.Namespace, "Request.Name", request.Name)
 	reqLogger.Info("Reconciling Hostcollector...")
 
-	hostCollector := &appdynamicsv1alpha1.Hostcollector{}
+	hostCollector := &appdynamicsv1alpha1.Clustercollector{}
 	err := r.client.Get(context.TODO(), request.NamespacedName, hostCollector)
-	reqLogger.Info("Retrieved host collector.", "Image", hostCollector.Spec.Image)
+	reqLogger.Info("Retrieved host collector.", "Image", hostCollector.Spec.HostCollector.Image)
 	if err != nil {
 		if errors.IsNotFound(err) {
 			// Return and don't requeue
@@ -165,7 +172,7 @@ func (r *ReconcileHostcollector) Reconcile(request reconcile.Request) (reconcile
 	return reconcile.Result{RequeueAfter: 60 * time.Second}, nil
 }
 
-func (r *ReconcileHostcollector) updateStatus(hostCollector *appdynamicsv1alpha1.Hostcollector) error {
+func (r *ReconcileHostcollector) updateStatus(hostCollector *appdynamicsv1alpha1.Clustercollector) error {
 	hostCollector.Status.LastUpdateTime = metav1.Now()
 
 	if errInstance := r.client.Update(context.TODO(), hostCollector); errInstance != nil {
@@ -182,25 +189,25 @@ func (r *ReconcileHostcollector) updateStatus(hostCollector *appdynamicsv1alpha1
 	return err
 }
 
-func (r *ReconcileHostcollector) hasBreakingChanges(hostCollector *appdynamicsv1alpha1.Hostcollector, existingDaemonSet *appsv1.DaemonSet) (bool, bool) {
+func (r *ReconcileHostcollector) hasBreakingChanges(hostCollector *appdynamicsv1alpha1.Clustercollector, existingDaemonSet *appsv1.DaemonSet) (bool, bool) {
 	fmt.Println("Checking for breaking changes...")
-	if hostCollector.Spec.Image != "" && existingDaemonSet.Spec.Template.Spec.Containers[0].Image != hostCollector.Spec.Image {
-		fmt.Printf("Image changed from has changed: %s	to	%s. Updating....\n", existingDaemonSet.Spec.Template.Spec.Containers[0].Image, hostCollector.Spec.Image)
-		existingDaemonSet.Spec.Template.Spec.Containers[0].Image = hostCollector.Spec.Image
+	if hostCollector.Spec.HostCollector.Image != "" && existingDaemonSet.Spec.Template.Spec.Containers[0].Image != hostCollector.Spec.HostCollector.Image {
+		fmt.Printf("Image changed from has changed: %s	to	%s. Updating....\n", existingDaemonSet.Spec.Template.Spec.Containers[0].Image, hostCollector.Spec.HostCollector.Image)
+		existingDaemonSet.Spec.Template.Spec.Containers[0].Image = hostCollector.Spec.HostCollector.Image
 		return false, true
 	}
 
 	return false, false
 }
 
-func (r *ReconcileHostcollector) newCollectorDaemonSet(hostCollector *appdynamicsv1alpha1.Hostcollector) *appsv1.DaemonSet {
+func (r *ReconcileHostcollector) newCollectorDaemonSet(hostCollector *appdynamicsv1alpha1.Clustercollector) *appsv1.DaemonSet {
 	trueVal := true
-	if hostCollector.Spec.Image == "" {
-		hostCollector.Spec.Image = "vikyath/host-collector:latest"
+	if hostCollector.Spec.HostCollector.Image == "" {
+		hostCollector.Spec.HostCollector.Image = "vikyath/host-collector:latest"
 	}
 
-	if hostCollector.Spec.ServiceAccountName == "" {
-		hostCollector.Spec.ServiceAccountName = "appdynamics-operator"
+	if hostCollector.Spec.HostCollector.ServiceAccountName == "" {
+		hostCollector.Spec.HostCollector.ServiceAccountName = "appdynamics-operator"
 	}
 
 	fmt.Printf("Building DaemonSet spec for image %s\n", hostCollector.Spec.Image)
@@ -211,7 +218,7 @@ func (r *ReconcileHostcollector) newCollectorDaemonSet(hostCollector *appdynamic
 			Kind:       "DaemonSet",
 		},
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      hostCollector.Name,
+			Name:      HOST_COLLECTOR_NAME,
 			Namespace: hostCollector.Namespace,
 		},
 		Spec: appsv1.DaemonSetSpec{
@@ -223,12 +230,12 @@ func (r *ReconcileHostcollector) newCollectorDaemonSet(hostCollector *appdynamic
 					Labels: ls,
 				},
 				Spec: corev1.PodSpec{
-					ServiceAccountName: hostCollector.Spec.ServiceAccountName,
+					ServiceAccountName: hostCollector.Spec.HostCollector.ServiceAccountName,
 					Containers: []corev1.Container{{
-						Image:           hostCollector.Spec.Image,
+						Image:           hostCollector.Spec.HostCollector.Image,
 						ImagePullPolicy: corev1.PullAlways,
 						Name:            "host-collector",
-						Resources:       hostCollector.Spec.Resources,
+						Resources:       hostCollector.Spec.HostCollector.Resources,
 						SecurityContext: &corev1.SecurityContext{
 							Privileged: &trueVal,
 						},
@@ -251,7 +258,12 @@ func (r *ReconcileHostcollector) newCollectorDaemonSet(hostCollector *appdynamic
 							Name:      "var-lib-docker",
 							MountPath: "/var/lib/docker/",
 							ReadOnly:  true,
-						}},
+						}, {
+							Name: "infraagent-config",
+							MountPath: "/opt/appdynamics/InfraAgent/agent.conf",
+							SubPath: "agent.conf",
+						},
+						},
 					}},
 					Volumes: []corev1.Volume{{
 						Name: "proc",
@@ -278,6 +290,14 @@ func (r *ReconcileHostcollector) newCollectorDaemonSet(hostCollector *appdynamic
 						VolumeSource: corev1.VolumeSource{
 							HostPath: &corev1.HostPathVolumeSource{Path: "/var/lib/docker/"},
 						},
+					},{
+						Name: "infraagent-config",
+						VolumeSource: corev1.VolumeSource{
+							ConfigMap: &corev1.ConfigMapVolumeSource{
+								LocalObjectReference: corev1.LocalObjectReference{
+									Name: INFRA_AGENT_CONFIG_NAME},
+							},
+						},
 					}},
 				},
 			},
@@ -292,8 +312,8 @@ func (r *ReconcileHostcollector) newCollectorDaemonSet(hostCollector *appdynamic
 	return ds
 }
 
-func saveOrUpdateHostCollectorSpecAnnotation(hostCollector *appdynamicsv1alpha1.Hostcollector, ds *appsv1.DaemonSet) {
-	jsonObj, e := json.Marshal(hostCollector)
+func saveOrUpdateHostCollectorSpecAnnotation(hostCollector *appdynamicsv1alpha1.Clustercollector, ds *appsv1.DaemonSet) {
+	jsonObj, e := json.Marshal(hostCollector.Spec.HostCollector)
 	if e != nil {
 		log.Error(e, "Unable to serialize the current spec", "hostCollector.Namespace", hostCollector.Namespace, "hostCollector.Name", hostCollector.Name)
 	} else {
@@ -304,7 +324,7 @@ func saveOrUpdateHostCollectorSpecAnnotation(hostCollector *appdynamicsv1alpha1.
 	}
 }
 
-func (r *ReconcileHostcollector) restartCollector(hostCollector *appdynamicsv1alpha1.Hostcollector) error {
+func (r *ReconcileHostcollector) restartCollector(hostCollector *appdynamicsv1alpha1.Clustercollector) error {
 	podList := &corev1.PodList{}
 	labelSelector := labels.SelectorFromSet(labelsForHostCollector(hostCollector))
 	listOps := &client.ListOptions{
@@ -324,6 +344,130 @@ func (r *ReconcileHostcollector) restartCollector(hostCollector *appdynamicsv1al
 	return nil
 }
 
-func labelsForHostCollector(hostCollector *appdynamicsv1alpha1.Hostcollector) map[string]string {
-	return map[string]string{"name": "hostCollector", "hostCollector_cr": hostCollector.Name}
+func labelsForHostCollector(hostCollector *appdynamicsv1alpha1.Clustercollector) map[string]string {
+	return map[string]string{"name": "hostCollector", "hostCollector_cr": HOST_COLLECTOR_NAME}
+}
+
+func (r *ReconcileHostcollector) ensureConfigMap(clusterCollector *appdynamicsv1alpha1.Clustercollector) error {
+	/*setContainerMonConfigDefaults(clusterCollector)
+	setServerMonConfigDefaults(clusterCollector)
+	*/
+	setInfraAgentConfigsDefaults(clusterCollector)
+	err := r.ensureContainerCollectorConfig(clusterCollector)
+	if err != nil {
+		return err
+	}
+	err = r.ensureInfraAgentConfig(clusterCollector)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func (r *ReconcileHostcollector) ensureContainerCollectorConfig(clusterCollector *appdynamicsv1alpha1.Clustercollector) error {
+
+	yml := fmt.Sprintf(`name: %s
+type: %s
+version: %s
+log-level: %s
+path: %s
+enabled: %t
+exporter-address: %s
+exporter-port: %d`, CONTAINER_COLLECTOR, TYPE_COLLECTOR, strings.Split(clusterCollector.Spec.HostCollector.Image,":")[1] , clusterCollector.Spec.LogLevel,
+		CONTAINER_COLLECTOR_PATH, true, clusterCollector.Spec.ExporterAddress, clusterCollector.Spec.ExporterPort)
+	cm := &corev1.ConfigMap{}
+	err := r.client.Get(context.TODO(), types.NamespacedName{Name: CONTAINER_CONFIG_NAME, Namespace: clusterCollector.Namespace}, cm)
+
+	create := err != nil && errors.IsNotFound(err)
+
+	if err != nil && !errors.IsNotFound(err) {
+		return fmt.Errorf("Unable to load clustermon configMap. %v", err)
+	}
+
+
+
+	cm.Name = CONTAINER_CONFIG_NAME
+	cm.Namespace = clusterCollector.Namespace
+	cm.Data = make(map[string]string)
+	cm.Data["clustermon.conf"] = string(yml)
+
+	if create {
+		e := r.client.Create(context.TODO(), cm)
+		if e != nil {
+			return fmt.Errorf("Unable to create clustermon configMap. %v", e)
+		}
+		fmt.Println("Agent Configmap created")
+	} else {
+		e := r.client.Update(context.TODO(), cm)
+		if e != nil {
+			return fmt.Errorf("Unable to update clustermon configMap. %v", e)
+		}
+		fmt.Println("Cluster collector Configmap updated")
+	}
+
+	return nil
+
+}
+
+func (r *ReconcileHostcollector) ensureInfraAgentConfig(clusterCollector *appdynamicsv1alpha1.Clustercollector) error {
+	errVal, controllerDns, port, sslEnabled := validateControllerUrl(clusterCollector.Spec.ControllerUrl)
+	if errVal != nil {
+		return errVal
+	}
+	portVal := strconv.Itoa(int(port))
+
+	yml := fmt.Sprintf(`name: %s 
+controller-host: %s
+controller-port: %s
+controller-account-name: %s
+controller-ssl-enabled: %s
+enabled: %t
+controller-access-key: %s
+controller-lib-socket-url: %s
+collector-lib-port: %s
+http-client-timeout: %d
+http-client-basic-auth-enabled: %t
+configuration-change-scan-period: %d
+configuration-stale-grace-period: %d
+debug-port: %s
+client-lib-send-url: %s
+client-lib-recv-url: %s
+log-level: %s
+debug-enabled: %t`, INFRA_AGENT_NAME, controllerDns, portVal, clusterCollector.Spec.Account, sslEnabled, true, clusterCollector.Spec.AccessSecret,
+		clusterCollector.Spec.SystemConfigs.CollectorLibSocketUrl, clusterCollector.Spec.SystemConfigs.CollectorLibPort,
+		clusterCollector.Spec.SystemConfigs.HttpClientTimeOut, clusterCollector.Spec.SystemConfigs.HttpBasicAuthEnabled,
+		clusterCollector.Spec.SystemConfigs.ConfigChangeScanPeriod, clusterCollector.Spec.SystemConfigs.ConfigStaleGracePeriod,
+		clusterCollector.Spec.SystemConfigs.DebugPort, clusterCollector.Spec.SystemConfigs.ClientLibSendUrl,
+		clusterCollector.Spec.SystemConfigs.ClientLibRecvUrl, clusterCollector.Spec.SystemConfigs.LogLevel, clusterCollector.Spec.SystemConfigs.DebugEnabled)
+	cm := &corev1.ConfigMap{}
+	err := r.client.Get(context.TODO(), types.NamespacedName{Name: INFRA_AGENT_CONFIG_NAME, Namespace: clusterCollector.Namespace}, cm)
+
+	create := err != nil && errors.IsNotFound(err)
+
+	if err != nil && !errors.IsNotFound(err) {
+		return fmt.Errorf("Unable to load infra agent configMap. %v", err)
+	}
+
+
+
+	cm.Name = INFRA_AGENT_CONFIG_NAME
+	cm.Namespace = clusterCollector.Namespace
+	cm.Data = make(map[string]string)
+	cm.Data["agent.conf"] = string(yml)
+
+	if create {
+		e := r.client.Create(context.TODO(), cm)
+		if e != nil {
+			return fmt.Errorf("Unable to create infra agent configMap. %v", e)
+		}
+		fmt.Println("Agent Configmap created")
+	} else {
+		e := r.client.Update(context.TODO(), cm)
+		if e != nil {
+			return fmt.Errorf("Unable to update infra agent configMap. %v", e)
+		}
+		fmt.Println("Infra Agent Configmap updated")
+	}
+
+	return nil
 }
